+++ R/drug_sensitivity.R
@@ -63,6 +63,8 @@
 #'  features from the file primary-screen-replicate-collapsed-treatment-info.csv 
 #'  
 #' - 21Q3: no change, no further releases are scheduled at this time.
+#' 
+#' - 21Q4: no change, no further releases are scheduled at this time.
 #'
 #' @docType data
 #'