+++ R/rnai.R
@@ -58,6 +58,8 @@
 #' - 21Q2: no change, no further releases are scheduled at this time.
 #'
 #' - 21Q3: no change, no further releases are scheduled at this time.
+#' 
+#' - 21Q4: no change, no further releases are scheduled at this time.
 #'
 #' @docType data
 #'